+++ src/ts-framework-tests/app.ts
@@ -1,4 +1,4 @@
-import {Application} from "../ts-framework/Application";
+import {Application} from "../ts-framework/Core";
 
 let app: Application = new Application(__dirname);
     app.start();
