+++ src/ts-framework/Core/Exception.ts
@@ -1,3 +1,2 @@
 export {Exception} from './Exception/Exception';
-export {Error} from './Exception/Error';
 export {AutoLoaderException} from './Exception/AutoLoaderException';
