--- src/ts-framework/Core/Exception/Error.ts
@@ -1,7 +1,0 @@
-export declare class Error
-{
-    public name: string;
-    public message: string;
-    public stack: string;
-    constructor(message?: string);
-}
