+++ src/ts-framework/Core/Exception/Exception.ts
@@ -1,4 +1,10 @@
-import {Error} from "./Error";
+export declare class Error
+{
+    public name: string;
+    public message: string;
+    public stack: string;
+    constructor(message?: string);
+}
 
 export class Exception extends Error
 {
