+++ src/ts-framework/Model.ts
@@ -2,4 +2,4 @@ export {DatabaseDriver} from "./Model/DatabaseDriver";
 export {IModel} from "./Model/IModel";
 export {Model} from "./Model/Model";
 export {ModelCollection} from "./Model/ModelCollection";
-export {ModelValidation} from "./Model/ModelValidation";
+export {IModelValidation} from "./Model/IModelValidation";
