+++ src/ts-framework/Model/IModelValidation.ts
